+++ src/main/java/com/soklet/internal/microhttp/ExpectationFailedException.java
@@ -1,0 +1,7 @@
+package com.soklet.internal.microhttp;
+
+class ExpectationFailedException extends RuntimeException {
+    ExpectationFailedException(String message) {
+        super(message);
+    }
+}